Move warm starting tests in separate file.
